requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
greenlet==3.1.1
idna==3.10
kagglehub==0.3.4
mysql-connector-python==9.1.0
numpy==2.1.3
packaging==24.2
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.36
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
